+++ packages/core/src/worker/brain/brief/index.ts
@@ -1,0 +1,6 @@
+export {
+  BrainBriefCompiler,
+  createBrainBriefCompiler,
+} from './compiler'
+
+export type { BrainBriefCompilerDeps } from './compiler'